finish config of shiro

+++ microservices/src/main/java/com/user/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 
-    User queryUserById(Integer id);
+    User queryUserById(Long id);
 
     User queryUserByEmail(String email);
 }
